extract verifyAppBridgeToken
