+++ space-plugins/nuxt-base/nuxt.config.ts
@@ -1,5 +1,9 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
 	devtools: { enabled: true },
-	ssr: false,
+	nitro: {
+		imports: {
+			dirs: ['./server/utils'],
+		},
+	},
 });
