+++ space-plugins/nuxt-base/server/api/_app_bridge.post.ts
@@ -1,22 +1,4 @@
-import jwt, { type VerifyCallback } from 'jsonwebtoken';
-import { DecodedToken, VerifyResponse } from '~/types/appBridge';
-
-export default defineEventHandler(async (event): Promise<VerifyResponse> => {
+export default defineEventHandler(async (event) => {
 	const { token } = await readBody(event);
-	try {
-		return { ok: true, result: await verifyToken(token, env('CLIENT_SECRET')) };
-	} catch (error) {
-		return { ok: false, error };
-	}
+	return verifyAppBridgeToken(token);
 });
-
-async function verifyToken(
-	token: string,
-	secret: string,
-): Promise<DecodedToken> {
-	return new Promise((resolve, reject) => {
-		const verifyCallback: VerifyCallback = (err, decoded) =>
-			err ? reject(err) : resolve(decoded as DecodedToken);
-		jwt.verify(token, secret, verifyCallback);
-	});
-}
